requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
openai==1.35.7
pydantic==2.5.2
python-dotenv==1.0.0
